fix test spec

+++ lib/convert_unit/base.rb
@@ -2,7 +2,7 @@ class Base
   require 'bigdecimal'
   attr_accessor :value, :unit
 
-  def initialize(value, unit, valid_units, precision=10)
+  def initialize(value, unit, valid_units, precision=16)
     raise TypeError, 'no implicit conversion of String into Integer' unless value.is_a? Numeric
     raise TypeError, 'Invalid Unit Type' unless valid_units.include?(unit.to_s.downcase)
     @value = BigDecimal.new value, precision
